# Implement comprehensive account performance optimizations