--- a/src/lib/trpc/router.ts
+++ b/src/lib/trpc/router.ts
@@ -1,4 +1,5 @@
 import { accountRoutes } from "$lib/trpc/routes/accounts";
+import { optimizedAccountsRoutes } from "$lib/trpc/routes/accounts-optimized";
 import { categoriesRoutes } from "$lib/trpc/routes/categories";
 import { payeeRoutes } from "$lib/trpc/routes/payees";
 import { transactionRoutes } from "$lib/trpc/routes/transactions";
@@ -9,6 +10,7 @@ import { scheduleRoutes } from "./routes/schedules";
 
 export const router = t.router({
   accountRoutes,
+  optimizedAccountsRoutes,
   categoriesRoutes,
   payeeRoutes,
   scheduleRoutes,