… drizzle/0000_secret_karma.sql
… drizzle/0001_slim_gunslinger.sql
… drizzle/db/sqlite.db
… drizzle/meta/0001_snapshot.json
… drizzle/meta/_journal.json
… src/lib/trpc/routes/transactions.ts
… src/lib/trpc/routes/views.ts
… src/lib/trpc/t.ts
… src/lib/utils/cache.ts
… src/lib/utils/performance.ts
… tests/integration/trpc/accounts-optimized-endpoints.test.ts
… tests/integration/trpc/accounts-optimized.test.ts
… tests/integration/trpc/categories.test.ts
… tests/integration/trpc/payees.test.ts
… tests/integration/trpc/transactions.test.ts
… tests/integration/trpc/views.test.ts
… tests/unit/middleware/input-sanitization.test.ts
… tests/unit/middleware/request-limits.test.ts
… tests/unit/utils/cache.test.ts
… tests/unit/utils/performance.test.ts
… vite.config.ts