## Summary

This PR introduces major performance improvements for account data loading through optimized tRPC endpoints with intelligent caching, pagination, and SQL-based aggregations.

### 🚀 Performance Improvements

- **65.7% faster** account summary loading compared to full data loading
- **Sub-millisecond** paginated transaction loading (0.77ms for 100 transactions)
- **SQL-based balance calculations** instead of JavaScript loops
- **Intelligent caching** with appropriate TTL based on data volatility
- **Server-side filtering** to reduce network payload

### 🏗️ New Optimized Endpoints

#### `optimizedAccountsRoutes.loadSummary`
- Loads account metadata + balance without transactions
- 65.7% performance improvement over full loading
- 5-minute caching for dashboard efficiency
- SQL aggregation for balance calculation

#### `optimizedAccountsRoutes.loadAllSummaries` 
- All account summaries in single optimized query
- Batch balance calculation for multiple accounts
- 2-minute caching for sidebar/navigation

#### `optimizedAccountsRoutes.loadTransactions`
- Paginated transaction loading (10-100 per page)
- Server-side search and date filtering
- Running balance calculation for chronological first pages
- Smart caching (1min for regular pages, no cache for searches)

#### `optimizedAccountsRoutes.loadRecentTransactions`
- Optimized for dashboard widgets
- Selective field loading (only id/name for payee/category)
- Configurable limit (5-20 transactions)
- 2-minute caching

#### `optimizedAccountsRoutes.getBalanceHistory`
- SQL-grouped balance history for charts
- Supports day/week/month grouping
- Date range filtering
- Performance monitoring integrated

### 🧪 Testing & Quality

- **23 comprehensive integration tests** (80 assertions)
- **Performance benchmarking** with realistic datasets  
- **Cache invalidation testing** to prevent stale data
- **Edge case handling** (empty accounts, large datasets, search queries)
- **Error handling** for non-existent accounts and validation errors

### 🔧 Technical Implementation

- **SQL aggregation** using Drizzle ORM for optimal database performance
- **Caching layer integration** with TTL-based expiration
- **Performance monitoring** using existing trackQuery infrastructure  
- **Input validation** with Zod schemas and security middleware
- **Type safety** maintained throughout the optimization layer

### 📊 Benchmark Results

```
Summary load: 0.36ms vs Full load: 1.04ms
Performance improvement: 65.7%
Paginated load time for 100 transactions: 0.77ms
```

### 🎯 Next Steps

This PR establishes the optimized backend foundation. Follow-up work will include:
- Frontend integration with TanStack Table for advanced data grids
- Dashboard widgets using the new recent transactions endpoint
- Charts integration using balance history endpoint
- Progressive migration from legacy endpoints

## Test Plan

- [x] All unit tests pass (`bun test tests/unit/`)
- [x] All integration tests pass (`bun test tests/integration/`)  
- [x] Performance benchmarks meet targets (>50% improvement)
- [x] Cache invalidation works correctly
- [x] Error handling covers edge cases
- [x] Large dataset performance validated

🤖 Generated with [Claude Code](https://claude.ai/code)